pyproject.toml
[project]
name = "bus-zeiterfassung"
version = "0.1.0"
description = "Simple self-hosted time tracker that fills a monthly Dienstzeitblatt Excel template and exports it to PDF."
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlmodel>=0.0.22",
    "jinja2>=3.1",
    "openpyxl>=3.1",
    "itsdangerous>=2.2",
    "pydantic-settings>=2.6",
    "python-multipart>=0.0.19",
    "argon2-cffi>=23.1",
]

[dependency-groups]
dev = [
    "ruff>=0.8",
    "mypy>=1.13",
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.28",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bus_zeiterfassung"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["openpyxl.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
